Bug-Fixes + generateLists() and writeToFile() commands
Fixed bug when reading from file that caused last character of data field to be cut off.

generateLists(): should be called at the very top of each command: calls a similar function to init that generates the appropriate lists, without checking for the file first (as init should take care of that)

writeToFile(): should be called at the very bottom of each function, writes current contents of the blockList to the file, in a format that appears to correctly be read back in when regenerating the lists from the file.